backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
pydantic-ai>=0.1.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
supabase>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
groq>=0.9.0
